Split out Locale, LocaleCodeHistory & LocaleQuerySet
